Added int parsing for constants comparisons develop branch. (#494)

+++ app/Http/Controllers/Api/V1/ActivityController.php
@@ -688,7 +688,7 @@ class ActivityController extends Controller
     public function getH5pResourceSettingsShared(Activity $activity)
     {
         // 3 is for indexing approved - see Project Model @indexing property
-        if ($activity->shared || ($activity->playlist->project->indexing === config('constants.indexing-approved'))) {
+        if ($activity->shared || ($activity->playlist->project->indexing === (int)config('constants.indexing-approved'))) {
             $h5p = App::make('LaravelH5p');
             $core = $h5p::$core;
             $settings = $h5p::get_editor();
